# implement a faster and simpler eachkmer