@bicycle1885 2017-05-19
The performance of kmer iterator improves for longer k-mers, too. However, for long k-mers, the cost of array access becomes dominant in `composition` as shown above. The following benchmarks do not access arrays so it is apparent that the performance has improved.

this branch:
```
julia> @benchmark count(_->true,each(DNAKmer{8}, seq))
BenchmarkTools.Trial:
  memory estimate:  80 bytes
  allocs estimate:  4
  --------------
  minimum time:     27.570 μs (0.00% GC)
  median time:      28.470 μs (0.00% GC)
  mean time:        30.255 μs (0.00% GC)
  maximum time:     118.795 μs (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     1

julia> @benchmark count(_->true,each(DNAKmer{16}, seq))
BenchmarkTools.Trial:
  memory estimate:  80 bytes
  allocs estimate:  4
  --------------
  minimum time:     27.635 μs (0.00% GC)
  median time:      30.678 μs (0.00% GC)
  mean time:        32.448 μs (0.00% GC)
  maximum time:     120.867 μs (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     1

julia> @benchmark count(_->true,each(DNAKmer{32}, seq))
BenchmarkTools.Trial:
  memory estimate:  80 bytes
  allocs estimate:  4
  --------------
  minimum time:     27.689 μs (0.00% GC)
  median time:      27.756 μs (0.00% GC)
  mean time:        30.375 μs (0.00% GC)
  maximum time:     265.790 μs (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     1
```

master:
```
julia> @benchmark count(_->true,each(DNAKmer{8}, seq))
BenchmarkTools.Trial:
  memory estimate:  80 bytes
  allocs estimate:  4
  --------------
  minimum time:     37.417 μs (0.00% GC)
  median time:      37.497 μs (0.00% GC)
  mean time:        40.360 μs (0.00% GC)
  maximum time:     171.461 μs (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     1

julia> @benchmark count(_->true,each(DNAKmer{16}, seq))
BenchmarkTools.Trial:
  memory estimate:  80 bytes
  allocs estimate:  4
  --------------
  minimum time:     35.936 μs (0.00% GC)
  median time:      36.722 μs (0.00% GC)
  mean time:        38.952 μs (0.00% GC)
  maximum time:     151.353 μs (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     1

julia> @benchmark count(_->true,each(DNAKmer{32}, seq))
BenchmarkTools.Trial:
  memory estimate:  80 bytes
  allocs estimate:  4
  --------------
  minimum time:     35.939 μs (0.00% GC)
  median time:      37.472 μs (0.00% GC)
  mean time:        38.220 μs (0.00% GC)
  maximum time:     158.646 μs (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     1

```